Same explicit cast fix for the alias-sync mechanism, for consistency.

+++ alphadog-v2-phase3c-certifier.js
@@ -476,7 +476,7 @@ async function reconcileHpBoardSubsetConstraints(pgClient, hpBatchId) {
       calibration_json = COALESCE(h.calibration_json::jsonb, '{}'::jsonb) || jsonb_build_object(
           'shared_threshold_alias_applied', true,
           'shared_threshold_alias_target', ${targetKeySide},
-          'hp_before_alias_sync', h.estimated_hit_probability_0_100,
+          'hp_before_alias_sync', h.estimated_hit_probability_0_100::numeric,
           'hp_after_alias_sync', (
             SELECT s.estimated_hit_probability_0_100 FROM score.hp_board_current s
             WHERE s.mlb_player_id = h.mlb_player_id AND s.hp_board_batch_id = h.hp_board_batch_id
